Abort on non-zero blastcmd exit code
Currently, Abricate will exit successfully regardless of the exit code
of the BLAST command. This is misleading as failures in the BLAST
command (eg. if it is killed by the OOM killer) will emit the message
"Found 0 genes in <file>", rather than indicating an error has ocurred.

This change checks the exit code of the executed command after close
is called. If it is non-zero, Abricate will terminate with a brief
message and the same exit code; allowing the calling shell to decide
what to do with the error.

This change uses a bash subshell instead of sh to make use of pipefail.
For anyone who wants to continue using sh I wave my hands in the air
and leave it as an exercise to anyone who wishes to inspect the exit
code of each piped command instead.